main: split argument parsing to separate file
We'll introduce much more environment variables in the next commit. Move
all arguments and env variable parsing to a new config.c file.
